// Frontend/client/src/app/models/product.model.ts
  export interface nannyss {
    FirstName: String;
    LastName: String;
    Region: String;
    image: String;
    Email: String;
    About: String;
    Age: Number;
    Phonenumber: Number;
  }
  
  
  // export interface serverResponse  {
  //   count: number;
  //   products: ProductModelServer[]
  // };